Add Cocoapods runner configuration
I ran into this error when trying to run the application:

```
The sandbox is not in sync with the Podfile.lock. Run 'pod install' or update your CocoaPods installation.
```

I overcame this error by running `pod install` in the `macos` and `ios` folders.
The files were updated with this content.