chore: Initial navigation
We use a `ShellRoute` for the main tabs; wallet and trade screen are wrapped by bottom navigation shell.

The Receive Screen was added to showcase a child screen that overlays the shell. We can also add child screens that are displayed within the shell, i.e. the bottom navigation will be visible.

Note: The settings screens were split by context, i.e.separate wallet and trade settings screen, because if we want to treat the settings screen as child screen the route would not be clearly defined if settings would be a child of both screens.