fix: Establish p2p connection between nodes
We've reworked the way that connections are established so that the
call to `keep_connected` returns as the connection is first
established. At the same time, `keep_connected` spawns a task which
tries to keep the connection alive.

This is a bit different to what we did in the `10101-poc`.

Also:

- Add short `tokio::sleep` to single-hop test before attempting to
connect nodes. This ensures that the attempt succeeds, but is only a
band-aid fix.

- Depend on `tracing-subscriber` to be able to configure the tests so
that we can see `tracing` logs.

- Tweak implementation of `lightning::Logger` for `TracingLogger`.
Still unsure if it works.