chore: common settings
Global settings with the wallet/trade settings highlighted upon navigating to the settings.

This allows us to have have a global settings screen, but have two routes for it so we can have it as child screen for wallet and trade. once implementing the settings screen we can decide to highlight the relevant settings from either screen upon navigating there.