[WIP] Debug why the channel doesn't open
We are able to connect the nodes and initiate the channel opening
after funding alice's wallet, but nothing appears to pop up on the
other side.

This commit has focused on improving the logs and adding an explicit
check so that we can tell when the channel is created. The check is
ugly atm, but we can clean that up.

Further notes for Philipp:

- You need to run `nigiri` and edit the constants in
`ln-dlc-node/test.rs` accordingly.

- You will need to create the `.ldk-data/alice/on_chain` and
`.ldk-data/bob/on_chain` directories ahead of time. Just let the test
run first and see how it fails for more context. This limitation of
the test setup is tracked in a couple of TODOs.