feat: regular announce node
from rust-lightning sample

Regularly broadcast our node_announcement. This is only required (or possible) if we have
some public channels, and is only useful if we have public listen address(es) to announce.
In a production environment, this should occur only after the announcement of new channels
to avoid churn in the global network graph.