chore: Log rust-lightning records to tracing
This is done through the `tracing_log` crate to use its dispatch_record function, as it is easier to convert a `lightning::Record` to a `log::Record` and convert it through that crate than to convert directly to a `tracing::Event`.

Global test subscriber was needed to capture traces from threads spawned by the rust-lightning libraries (e.g lightning-background-processor spawns its own threads, which we cannot install the local subscriber in).

I chose to use tracing values instead of a formatted log message as they are searchable through tracing log ingestors. This is mostly stylistic though and could go back to how it was before

Co-authored-by: Restioson <restiosondev@gmail.com>
Co-authored-by: Lucas Soriano <lucas@comit.network>