fix: make it work with keyboard on iPhone
With these changes the mobile keyboard correctly pushes the modal sheet up and one can close the keyboard on tapping into the sheet.
Note that the sheet content currently requires a fixed height, otherwise it consumes the complete height of the screen.
Eventually it would be good to achieve that the children size just define the size of the sheet, but for now the fixed size is good enough.