fix: remove splash screen after init
The problem is that on android I keep running into an error when initializing the rust logs - and we don't remove the splash screen if tere is an error.
The fact that we run into initialization error should not result in the app being stuck on the splashscreen.
With removing the splashscreen in the `finally` block we ensure that it is removed either way.
The `ready` variable is not needed then.
Additional strategies to tell the user that something went wrong might be needed.