fix: Invoice amount can not be higher than 10% of the total channel value
Just noticed that our single hop test is not working and found that magical piece of configuration. In our test we created a channel from Alice to Bob with 30.000 sats. Due to the max_inbound_htlc_value_in_flight_percent_of_channel config I was only able to send 3.000 sats at once. It seamed odd to me that I am restricted by that, imho that limit should be far higher, but at the same time I don't see the risks related to that.