chore: format first, then clippy, then tests
Adding a dependency on the jobs will lead to sequential execution of the ci jobs:
1. formatting
2. clippy
3. tests

The list is ordered by time complexity, i.e. the fastest one is first. This is done because if  formatting fails, developers will push again to make ci happy. However, the first push and the second push both will make ci run `clippy` and `tests` if there is no dependency.
Hence, I added this dependency so that devs have to ensure proper formatting first before clippy or the tests are being run.
This will free up some ci resources which makes other PRs run faster.