feat(ln-dlc-node): Support creation of DLC channel
This patch includes too many things that I have not managed to split
into smaller patches.

Namely:

- Add new test to prove that we can open a DLC channel.
- Start refactoring the "production" code in `ln-dlc-node`.
- Rework all the tests in `ln-dlc-node`. The objective was to simplify
and better understand why things work. This led to fixing a couple of
bugs, and it made the tests faster and easier to read.
- Remove `system_tests` feature in favour of `#[ignore]` attributes.

The DLC functionality still needs to be tested more e.g. can we make a
payment after opening a DLC channel.