feat: Expose get_fake_scid endpoint on coordinator
Exposing that communication requirement through http is the simplest way of getting the fake channel id from the coordinator, however, it also adds an additional dependency to rocket, which might be an overkill.

We might want to evaluate exchanging the fake channel id through a custom message over the p2p lightning connection or find a way to not require that communication at all.