chore: Increase max inbound htlc value of in flight payments
Upon creation of a just-in-time channel we use the max of either the `LIQUIDITY_ROUTING_FEE_MILLIONTHS` constant (10_000) or twice the amount of the in-flight payment to define the channel capacity.

In case we do not use the constant, the payment will fail with the current configuration on main, as the `max_inbound_htlc_value_in_flight_percent_of_channel` of the just-in-time-channel rejects pending HTLCs greater than 10% of the total channel capacity.

So, given our current calculation of the channel capacity, this value has to be at least 50%. However, we are setting it to 100% though - see reasoning in the comments.