feat: Calculate cfd fees
This adds the correct fees to the trading confirmation screen. However, these fees should probably be already shown on the trade screen. As you these fees are not that small.

This change uses the rust-dlc implementation for its calculation as this will implicitly be used when we are proposing a sub channel.