feat: introduce database for mobile app
We make use of sqlite+diesel for the mobile app. This allows us to write higher order queries.
Note the file changes in
- mobile/ios/Runner.xcodeproj, and
- mobile/native/Runner.xcodeproj
We need to add the LD flag `"-lsqlite3" as otherwise on MacOS neither the iOS App nor native app for MacOs will start up.

This commit also includes a sample model including a simple update and test. I thought I keep it for the time being so that we know things work until we have a useful model. I'll drop these changes in a follow-up commit.