fix: Announce the address the node is listening on
Announcing our node hard-coded to 127.0.0.1 resulted in the lnd not being able to reconnect to our coordinator if he got restarted. (Note, lnd runs in docker and the coordinator not). Thus, the channel changes to `active: false` and can not be used for payments anymore. Note, that even after this change, lnd will need some time (a few seconds) to process the coordinator being online again and set the channel to active.

This change will take the provided address to be used for node announcement. Note, that binding the ln_dlc node to `0.0.0.0` will also publish the ip `0.0.0.0` in the node announcement. This IP however, will not be useful for lnd to reconnect. So the actual local or public ip address has to be provided in order for the lnd node to be able to reconnect to the coordinator after a restart.

Initially I thought this node announcement is not needed at all, but it looks like it is required for the creation of the just-in-time channel. If the mobile app does not announce it's node, the coordinator will fail to create a just-in-time channel with the error `Unable to decode our hop data`. I believe this is because the coordinator does not store the connection data of the app, unless there is an open channel.

Interestingly the coordinator does not need to announce his node in order for lnd to be able to reconnect after a coordinator restarts. I believe this is because in our faucet setup the lnd initiates the connection and channel and thus persists the connection data of the coordinator.