feat: Accept configuration in mobile app
Leveraging `--dart-define` this change passes the configuration set from flutter to the rust backend. Note, unfortunately we can not directly access the environment variables from rust as it is done in dart.

Alternatively this could have been implemented with a config file, that is being read from, but I decided against it as it feels less flexible than a command line argument. see also https://suragch.medium.com/reading-configuration-files-and-settings-in-flutter-and-dart-cc14bbd93698