fix: Nest the `share_invoice` screen in `create_invoice` screen
This achieves that when clicking the back button in the `share_invoice` screen we navigate back to the `create_invoice` screen.
Since the routes in `go_router` are declarative, there is not easy way to create a "dynamic" wizard, where we navigae back to the "previous" screen. If we want this to be dynamic with an index we would have to use one screen and a navigator within that screen that changes the content of the screen.
For now, nesting the screen is a simple solution to achieve the behavior we want.