chore: Minimise need to depend on rust-dlc in other crates
This dependency should ideally only be known by core crates such as
`ln-dlc-node`.

We define a method `get_new_address` on `Node` so that we can control
how we call the underlying on-chain wallet, rather than letting
consumers call the wallet through a trait that is designed for a
completely different purpose.

We also turn the generic `get_dlcs` API into the more specific
`get_confirmed_dlcs`, as this is the only thing that is currently
needed. This way we can avoid consumers needing to understand types
defined by `rust-dlc`. The `get_confirmed_dlcs` API and the matching
`Dlc` type should evolve (and perhaps even disappear) as we meet new
requirements.