chore: Add Clippy lint workspace configuration
We've used a similar pattern in past projects.

One thing that has changed is that we use
`[target.'cfg(feature = "cargo-clippy")']` instead of `[build]`. This
allows us to enforce this lint configuration locally and not just on
CI.

Using `rustflags` does mean that devs might encounter unexpected full
rebuilds when switching between IDE and command line when using `cargo
clippy`. This is a known issue[1]. For emacs, I just solved this by
setting a different Cargo `target-directory` to the one used by
default. A similar approach should work for other IDEs.

[1]: https://github.com/rust-lang/cargo/issues/8716.