fix: Rename the `Order` `status` to `state`
The enum that identifies this field is called `OrderState`, the field should be called `state`.

Note that this rename was driven by a problem with formatting + diesel migrations:

The `status` field in the `CREATE TABLE` SQL query was change to `STATUS` when formatting because that seems to be an SQL keyword - but it appears that the field names are case sensitive, so Diesel ran into problem once re-running the migrations!
Renaming the field to `state` throughout (as it should be anyway) fixes this.