feat: Persist and update the order
This drives the order lifecycle for the *happy path* forward.
We don't handle setting the order into a failed state yet.

We save the order in the database upon submission.
The order is updated upon trading (which is currently still triggered straight after order submission; this will change once orderbook events are being consumed).
Once we get the confirmed DLC (i.e. the trade is done) we load the order that was set to `Filling` upon triggering the trade.
We currently assume that there is always only one order in `Filling`. If there is no order in `Filling` there is not trade ongoing.

Note that with the order now being available once the position is being updated, the position reflects the correct values.