chore: Use docker volume instead of file system for postgres
I was stumbling over this error when starting the maker.

```
2023-03-08 10:12:46 ERROR r2d2: connection to server at "localhost" (::1), port 5432 failed: FATAL:  database "maker" does not exist
```

It appears I had an old postgresql database state on the filesystem which had to be cleared so that the init script will be executed again.

Although this is a workable solution, it's nicer to mount the postgres data into a docker volume instead the file system. This way a simple `docker-compse down -v` will also clear the postgres volume.