feat: Handle order failure scenarios when starting trade execution
This handles handles order failure for problems that happen *before* the know anything about the DLC state from the DLC manager.
If we fail to update state of an order in the database, request trade execution with the coordinator or have trouble with our lightning node then the order will be set to `Failed` with a failure reason.

For the MVP we don't implement retry.

Note that failure scenarios due to the DLC protocol execution are not yet handle with this commit.