WIP: Publish order filling event / testing
Setup gets a bit complex at this point as the maker is now also included in the test. We need to find a way to do this automated as there a so many moving parts currently.

Manuell testing failed with either the taker or the maker channel getting closed during dlc channel proposal. see message below.

```
2023-03-10 19:12:58 DEBUG lightning::ln::peer_handler: Got Err message from 02b103838b4fc38a423342e2d187de6da76dde13e7a0271c8247e19c91027140f7: Channel closed because of an exception: Funding transaction was un-confirmed. Locked at 1 confs, now have 0 confs.
2023-03-10 19:12:58 ERROR lightning::ln::channelmanager: Force-closing channel dbab3910068bf81fd7c0148cc5ef48686528036493e0a43324d54c7b4839d43b
2023-03-10 19:12:58 DEBUG lightning::ln::channelmanager: Finishing force-closure of channel with 0 HTLCs to fail
``