feat: Close position
The coordinator offers a collaborative dlc channel close after receiving an order to an already existing position.

Note, we do not check if this is really a reverse order, yet, but rather simply imply that the order is intended to close the position.

This PR is still using the mock data, that need to be wired with actual in eventually.