refactor: Start the lifecycle of the order after submission to orderbook
If we succeed to submit the order we assign the `Uuid` returned from the orderbook to the order and safe it in the database.
If the fail to submit the order we assign our own `Uuid` to the order and safe it in the database.
Given that the lifecycle now starts with the result of the submission to the orderbook we don't need an `Initial` state anymore.