refactor: Move trade params to `coordinator-commons`
The `TradeParams` are specific to the coordinator, it makes sense to move them to a `coordinator-commons` crate.
This has the advantage that we will avoid cyclic dependencies, because we can use `orderbook-commons` to reference the `FilledWith` in `coordinator-commons` and are able to use the global common types of `trade` in other crates.