fix(coordinator): Build payout function pieces correctly
We have to skip the "upper" payout function piece if the "middle"
payout function piece already reaches the `BTCUSD_MAX_PRICE`. For one,
there is no need for another payout function piece, but if we do add
one and it uses the same point for both ends of the piece `rust-dlc`
blows up.