wip: make trading flow work
We had a problem with the order_id:
the orderbook can't come up with the order-id because we need to store it in the db on the client side before getting the id from the orderbook.
This is why we let the client come up with the order id and not the orderbook.

On the way we also solved a few more problems and added some more error handling.