refactor: Move message processing from `lndlcnode` into the client
The client should be able to control the message flow (i.e. intercept messages, react on them, ...), so we moved this from the Node into the client.
This allows us to react to e.g. the `Finalized` message so we can react on something happening rather than having to constantly poll the state.