chore: Use DLC channel terminology in logs
In our domain, it makes more sense to talk about a DLC channel as it
is more descriptive: it's the channel that holds the DLC, as opposed
to the Lightning channel.

The "sub-channel" terminology comes from `rust-dlc` where they have to
distinguish between vanilla DLC channels (where no Lightning channel
is involved) and sub-channels (where the DLC channel is a sibling of
the Lightning channel).