feat: Disable close button in position `Closing` state
Note: We should additionally not allow pressing the button *right* after it was pressed in the user-interface, because this logic does not prevent the user from pressing it twice.
Note that, eventually, we will also have to track if the order for closing the position failed and re-set the state to `Open`.