fix: Enable scrolling position/order list by adding bottom spacer
Due to the Buy/Sell floating action buttons in the middle some information may be hidden.
By adding a spacer we allow the user to scroll to move the complete tile into focus.
The spacer at the bottom is always added; since it is at the bottom of the list it does not do any harm.