Harmonize order state handling
Harmonize how the order state is updated.
Initially we opted for not exposing `Rejected` because we would exit with an error in the order submission dialog; however, it is not guaranteed that the user will see this error.
To keep things simple we map the internal `Rejected` state to `Failed` in the user interface, and update the order accordingly. This guarantees that the user will know that the order failed.