chore(coordinator): Add proptest for payout function
One can play around with the test by choosing different fixed values
of `rounding_mod` to verify that for most values we can run into a bug
in `rust-dlc`.

Whilst we shouldn't test `rust-dlc` (our dependencies in general)
here, we could add a test _like_ this one to `rust-dlc` once the bug
is fixed to probabilistically prove that the fix was sufficient.