fix(event_handler): Bubble up certain errors instead of panicking
The `unwrap`s removed were not safe to use as they corresponded to
function calls that could normally result in an error which should be
handled gracefully.

On the other hand, we left a few `unwrap`s that were deemed valid.
Specifically, it is okay to `unwrap` mutex locks as one cannot recover
from a poisoned mutex anyway. The application should just be restarted
in such a situation. That's not _yet_ handled in this patch, as the
panic occurs in a task which does not lead to aborting the program.

Co-authored-by: Richard Holzeis <richard@holzeis.me>