chore(ln-dlc-node): Set rounding_mod for test based on n_cets
We're used to thinking about the discretisation of the payout curve in
terms of number of CETs. We also know that the more CETs, the more
time it takes to complete what we used to call "contract setup".

We can therefore use the `n_cets` variable to control how fast these
tests take to complete the "contract setup" phase.