chore: Update to latest rust-dlc
Other than breaking changes to certain `rust-dlc` APIs, the main
hurdle was to fix errors on `ln-dlc-node` tests that built on first
opening a DLC channel.

To do so we had to manually instruct the channel offerer to process
the `Finalize` message. Before the dependency update this wasn't
necessary, but now one must finish the channel open protocol before
doing other things with the channel (such as closing it
collaboratively), which makes sense.