fix(mobile): Handle opening and closing of DLC channels separately
A closed DLC channel results in a `Contract::Closed` state, as opposed
to a `Contract::Confirmed` state. As such, we were unable to find a
`Confirmed` DLC after handling the `CloseFinalize` sub-channel
message, which led to a bug[1].

But it turns out that if we are reacting to a DLC channel being
closed, we already know that the position should change in a different
manner (it should simply be deleted). We therefore separate the
handling of opening and closing DLC channels on the way to fixing the
bug.

[1]: https://github.com/get10101/10101/issues/359.