fix: Reset lnd mission control on send payment
lnd is keeping a heuristic for routing attempts based on past payment attempts. This information is kept in the mission control and seems to block any subsequent payment to a node that has previously failed. If it is the very first payment that fails I guess this heuristic is even heavier opinionated.

https://docs.lightning.engineering/lightning-network-tools/lnd/payments#mission-control

In order to bypass this mission control, this change simply resets the mission control on every payment.