chore: Modify and redesign keep_connected API
It is only used by the app, so it is simpler to move it to the
`mobile` crate, rather than trying to define something generically
useful.

With the move we notice that the app only needs a never-ending future
that will constantly attempt to establish and reestablish the
connection with the coordinator when necessary. With that insight, we
can remove the part that was ensuring that the method would only
return once the initial connection had been established, simplifying
the code.

At the same time, this also allows us to redesign `Node::connect` and
remove `Node::connect_to_peer`, since they are used for the same
purpose.