chore(ci): Time out run-tests-ln-dlc-node job after 30 minutes
We have seen these tests hang for hours, which is wasteful and
annoying as we don't get feedback fast enough unless we actively check
the Actions tab.

Choosing 30 minutes as a timeout seems sensible as no successful run
has ever gone over 15 minutes and most of them are much faster than
that.