Add validation for channel capacity
Makes sure that the user can only trade within the bounds of the channel capacity.
The user can only trade a margin of: channel_capacity - total_reserve - counterparty_margin.

In order to enable users to trade that have an unbalanced channel we display an info dialog that recommends that only 50% of the channel should be filled up.