Fixed margin instead of quantity
This makes it easier for the user to trade, because the pre-filled value is usable straight away (unless we have insufficient balance or insufficient capacity).

For simplicity we opt to set the default margin to the minimum trade-able margin, which is `1000` sats.
In the future it would make sense to set it to the maximum trade-able amount, but that is more complicated because it depends on knowing the usable balance which we currently don't have when initializing the trade values.

In order to keep the margin stable, when changing the price we re-calculate the quantity instead of the margin.
When changing the leverage we still change the margin because that makes more sense in my head (leverage reduces the margin, not increases the quantity - but some users might disagree...).