Send screen overhaul
- Decode invoice
- Form validation
- Success dialog

Taking shortcuts was hard unless we accept an experience where the user does not get *any* feedback after pressing the button - which I think is unacceptable.
Validation requires decoding (otherwise there is not much to validate), and showing the user success or failure requires proper state transitions so I went the mile and added it.

Note: Fixed an issue where we did not set the `payee_pubkey` in our invoices on the way - this should make handling invoices more performant.