fix(mobile): Handle order state transitions a bit better
We've moved the state-check before updating the state of an order
inside of the actual DB transaction to ensure atomicity. This lack of
atomicity was very likely the cause of
https://github.com/get10101/10101/issues/438.

We've enhanced the state-check and converted it into the
`OrderState::next_state` method which determines what is the correct
update in all cases.

Moving the state-check into the DB transactions did cause them to be
long enough that we would frequently run into pesky `database is
locked` errors. To fix this, we have re-configured the
`MAX_DB_POOL_SIZE` to 1. This might seem very low, but it has
eradicated the `database is locked` errors and the performance is
fine given the simplicity of the app. I don't expect this change to be
permanent, but it appears to be good enough for now.