fix(ln-dlc-node): Do not wrongly unconfirm funding TX
With the version of `rust-lightning` prior to this patch[1], after
opening a DLC channel we could wrongly unconfirm the LN funding
transaction if it had already been confirmed.

We would look for the glue transaction on-chain, not find it and then
report back to the `ChannelManager` that the _funding_ transaction was
unconfirmed. This led to force-closing the LN channel unnecessarily.

[1]:
https://github.com/p2pderivatives/rust-lightning/commit/a33cc845bc05eb0bb3e9ba8f11789ba4ed44f5b1.