feat: Delay payment for the user to open 10101
In cases the user will onboard to 10101 using another mobile app. 10101 will be closed at the time the payment is sent. Without delaying payment processing before the user opens 10101 the payment will always fail.

This is a quick fix to solve that issue, by simply waiting for 10 seconds for the 10101 receiver to come online.

Eventually we should implement some logic that is keeping 10101 online in the background for some time to cater for that scenario more nicely.

Also note, that this fix assumes that the sending party will remain online while the payment is sent.