feat(coordinator): Stay connected to public channel peers
We want to stay connected to our public channel peers so that our
channels can be usable and payments can be routed through them at all
times.

This approach should work for all public channel peers who are
listening on clearnet. If we want to reconnect to any peers who use
Tor, we will need to use a Tor proxy[1].

[1]: https://github.com/TonyGiorgio/ldk-sample-tor.