chore(ln-dlc-node): Remove misplaced TODO
The apparent requirement is that the routing node must provide a node
announcement to the funding LND node _after_ creating the channel
between the two and _before_ the payment is made.

The code next to the removed TODO is about periodic node
announcements, which are recommended by the LDK devs[1]. The TODO is
therefore a bit confusing there.

[1]:
https://discord.com/channels/915026692102316113/978829624635195422/1100452596726116383.