chore: Change data directory on phone
On iOS the ApplicationSupportDirectory is never exposed to the user and can not be backed up. Contrary the ApplicationDocumentsDirectory is user generated data, that can be backed up. I think this still not perfect as we have both kinds of data, but beeing able to access the data dir on the phone has IMHO several benefits. Especially during the beta phase.

Note: Seed is still stored in the original location.

Co-authored-by: Richard Holzeis <richard@holzeis.me>