Change intervals for regtest
Introduce different intervals for wallet and node announcements on regtest (the
same values that we used to have at the beginning).

Without these, quickly testing on regtest became unfeasible due to errors in funding.

- Lack of wallet sync causes `insufficient_balance` issues right after funding.
- Lack of node announcement causes `no_route` issues.

Note: There are better ways to handle these errors, but we don't have time for that
right now (it's not an issue in production).

Potential long-term fixes:

- Fixing wallet sync:
    - add API to manually sync it after funding

- Fixing broadcasting node:
    - delay the initial node announcement