Remove tip cache in favour of blocking impl
This is tracked as separate commit to show that we tried this to understand if the tip implementation is related to the `panic` error we see in the tests:

```
thread 'tokio-runtime-worker' panicked at 'Untracked confirmation of funding transaction. Please ensure none of your channels had been created with LDK prior to version 0.0.113!', /Users/dakami/CoBloX/repos/10101/crates/lightning-transaction-sync/src/esplora.rs:391:17
```

Fetching the tip sync did not help mitigating the panic, but we keep the sync implementation because it is likely more error prone because it has less unknowns (and fetching height and blockhash should be very cheap).