fix(ci): Run ln-dlc-node tests sequentially
I think we can remove the `-j 1` flag, because I introduced it to
achieve the same thing and it clearly wasn't working.

We must go down to `--test-threads=1` if we want to ensure that no two
tests are run in parallel. Running tests in parallel at all seems to
pretty consistently cause either test failures or non-terminating
tests. Perhaps this change might fix
https://github.com/get10101/10101/issues/355.