feat(ln-dlc-node): Use on-chain address for shutdown and destination
=== Collaborative close of LN channel ===

Before, the collab close transaction paid to an address owned by LDK,
which had to be confirmed and then spent to the on-chain wallet.

After overriding `get_shutdown_scriptpubkey`, the close transaction
pays directly to the on-chain wallet. This means that there is no risk
of losing funds as soon as the close transaction is on-chain if we
remember the seed, whereas before we could lose funds whilst waiting
for the "claim transaction".

=== Force-close of LN channel ===

The effect of overriding `get_destination_script` to point to an
on-chain wallet address is less clear to me, because the commitment
transaction must still be processed by LDK to produce a
`SpendableOutputs` event with a
`SpendableOutputDescriptor::DelayedPaymentOutput` which can then be
spent to an on-chain wallet address (after waiting for the
corresponding delay).

=== Other changes ===

The tests were fixed to pass after the new changes were introduced. At
the same time we decided to refactor them with the goal of making them
clearer and simpler. These two tests diverged enough that it made
sense to remove the common helper function.

=== Caveat ===

The trait method implementations modified include a few `expect`s
because the trait methods are defined as infallible. We copied the
code from `ldk-node`, so at least that suggests that they are okay
with these `expect`s.