feat(ln-dlc-node): Let coordinator force-close LN-DLC channel
This patch finally adds a `force_close_ln_dlc_channel` test, which
verifies that the party effecting the force-close of the LN-DLC
channel can get their coins back if they:

- Manually call `force_close_channel`, publishing the split
transaction.

- Wait for 288 confirmations on the split transaction.

- Manually call `finalize_force_close_ln_dlc_channel`, publishing
the buffer, glue and commitment transactions.

- Wait for 144 confirmations on the commitment transaction, so that
the delayed output on it becomes spendable.

- Wait for a further 144 confirmations, so that the CET can spend from
the buffer transaction.

For the time being only the coordinator is able to force-close a
channel by first using the DELETE /api/admin/channels/:channel_id
endpoint and later the new DELETE
/api/admin/channels/finalize_force_close/:channel_id endpoint.

=== Limitations ===

- This work does not yet verify that the party who did not trigger the
force-close gets all their coins back. Specifically we are aware that
there is a bug with the `StaticPaymentOutput` on the commitment
transaction not being spendable.

- The updates to the `rust-dlc` and `rust-lightning` dependencies
represent very hacky changes that "just work".

- The `force_close_ln_dlc_channel` test should not assert on a magic
number, as this is hard to understand and brittle.