feat: Replace faucet button with direct channel to coordinator
This is to ensure that we have a more reliable way to open a channel
between coordinator and app on regtest.

The user still clicks on the newly introduced `Pay the invoice with
10101 faucet` button, but under the hood we just open a channel
directly and push 100k sats to the app. We could try to read what the
user asked for when creating the invoice, but that's more involved and
we just wanted a simple fix.