Change `sync` to use a multi-threaded runtime builder
We copied the code that uses `new_current_thread` from LDK wallet.
In my understanding we need this because `SledTree` does not implement `Send`, so we need a non-send thread.

However, the tokio documentation states that one should use `LocalSet` when using `new_current_thread` - which is not done in the code copied over from `ldk-node`.
 The tokio docs further state that one can use `new_multi_thread` when having the `rt-multi-thread` feature activated - which we have.