chore: Make process_pending_dlc_actions sync
The function `process_pending_dlc_actions` certainly does not benefit
from being async.

Unfortunately, the underlying `rust-dlc` API `process_actions` varies
in how expensive it is so it's not obvious whether we should allocate
a dedicated thread for this work or not. For the time being we follow
the pattern of having an inexpensive, long-running async task which
routinely calls `spawn_blocking` to do the blocking (sometimes
expensive) work in a short-lived, blocking thread.