fix(ln-dlc-node): Use correct fee rate in JIT channel tests
These tests were using a constant which is no longer relevant since we
use the esplora client to pull the fee rate estimates based on the
test blockchain. This was causing the now-renamed
`just_in_time_channel_works_with_correct_fee` test to fail.

The only sane way that I could find to fix this was to refactor
everything surrounding the JIT channel tests. This allows us to get
rid of the `TestPath` enum, which was making the tests hard to read.