fix(ln-dlc-node): Locate claim TX in force close test
The fee calculations have changed (probably since the introduction of
the esplora client), so this test was failing.

Since the test is not really about verifying exact fees, it's
sufficient to find the _only_ transaction that pays to the payee's
on-chain wallet. By verifying that the transaction is incoming we
ensure that is the claim transaction (in the context of the test).

These assertions are not perfect, but I think they are good enough.