Add sync method - Problems with async started
Seems like we can't use the async interfaces :(

I went on a different branch and tried to hand-craft an async version, but it
didn't get me anywhere.

The next thing is to drop the revert and go with sync bdk until then.

The advantage of cached WalletInfo with accessing should still give us pretty
good results...