chore(ln-dlc-node): Add load test
The motivation to add this test is to check that a hosted coordinator
is able to trade repeatedly with an app. In this case we simply
simulate the app rather than using the `mobile/native` crate.

Notably, we have to interact with the coordinator's HTTP API. To do so
we have to pretend like we got a match from the orderbook. We could
involve the orderbook too, but it isn't strictly necessary at this
stage.

The test in its current state is much slower than it needs to be. We
should work towards replacing the hard-coded sleeps with calls to
`wait_until`, wherever possible.

We've added a `load_tests` flag because we probably should refrain
from running this type of test as frequently as the rest of the
`ln-dlc-node` tests. Currently it relies on the public regtest
`coordinator` being up an running, and sufficiently funded. Eventually
we might want to run this kind of test, for example, before every
release.

Co-authored-by: Richard Holzeis <richard@holzeis.me>